Update endpoints, remove deprecated routes/db connection
